+++ test/twemoji_test.rb
@@ -6,7 +6,30 @@ class TwemojiTest < Minitest::Test
   end
 
   def test_number_of_emojis
-    assert_equal 874, Twemoji::CODES.size
+    assert_equal 1661, Twemoji.codes.size
+  end
+
+  def test_twemoji_png_and_svg_not_loaded_by_default
+    assert_raises NameError do
+      Twemoji::PNG
+      Twemoji::SVG
+    end
+  end
+
+  def test_twemoji_png
+    require "twemoji/png"
+    assert_raises NameError do
+      Twemoji::PNG
+    end
+    assert_equal 1661, Twemoji.png.size
+  end
+
+  def test_twemoji_png
+    require "twemoji/svg"
+    assert_raises NameError do
+      Twemoji::SVG
+    end
+    assert_equal 1661, Twemoji.svg.size
   end
 
   def test_finder_methods_cannot_find_by_more_than_one
@@ -46,64 +69,64 @@ class TwemojiTest < Minitest::Test
   end
 
   def test_parse_plus_one
-    expected = %(<img draggable="false" title=":+1:" alt="👍" src="https://twemoji.maxcdn.com/16x16/1f44d.png" class="emoji">)
+    expected = %(<img draggable="false" title=":+1:" alt="👍" src="https://twemoji.maxcdn.com/2/svg/1f44d.svg" class="emoji">)
 
     assert_equal expected, Twemoji.parse(":+1:")
   end
 
+  def test_parse_minus_one
+    expected = %(<img draggable="false" title=":-1:" alt="👎" src="https://twemoji.maxcdn.com/2/svg/1f44e.svg" class="emoji">)
+
+    assert_equal expected, Twemoji.parse(":-1:")
+  end
+
   def test_parse_html_string
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="emoji">!)
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="emoji">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!")
   end
 
   def test_parse_document
     doc  = Nokogiri::HTML::DocumentFragment.parse("<p>I like chocolate :heart_eyes:!</p>")
-    expected = '<p>I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="emoji">!</p>'
+    expected = '<p>I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="emoji">!</p>'
 
     assert_equal expected, Twemoji.parse(doc).to_html
   end
 
   def test_parse_option_asset_root
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://emoji.bestcdn.com/16x16/1f60d.png" class="emoji">!)
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://emoji.bestcdn.com/svg/1f60d.svg" class="emoji">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", asset_root: 'https://emoji.bestcdn.com')
   end
 
-  def test_parse_option_file_ext_svg
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/svg/1f60d.svg" class="emoji">!)
+  def test_parse_option_file_ext_png
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/72x72/1f60d.png" class="emoji">!)
 
-    assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", file_ext: '.svg')
+    assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", file_ext: 'png')
   end
 
   def test_parse_option_unsupport_file_ext_gif
     exception = assert_raises RuntimeError do
-      Twemoji.parse("I like chocolate :heart_eyes:!", file_ext: '.gif')
+      Twemoji.parse("I like chocolate :heart_eyes:!", file_ext: 'gif')
     end
 
-    assert_equal "Unsupported file extension: .gif", exception.message
-  end
-
-  def test_parse_option_image_size
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/72x72/1f60d.png" class="emoji">!)
-
-    assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", file_ext: ".png", image_size: "72x72")
+    assert_equal "Unsupported file extension: gif", exception.message
   end
 
   def test_parse_option_class_name
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="twemoji">!)
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="twemoji">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", class_name: 'twemoji')
   end
 
   def test_parse_option_single_img_attr
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="twemoji" style="height: 1.3em;">!)
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="twemoji" style="height: 1.3em;">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", class_name: 'twemoji', img_attrs: { style: 'height: 1.3em;' })
   end
 
   def test_parse_option_multiple_img_attrs
-    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="twemoji" style="height: 1.3em;" width="20">!)
+    expected = %(I like chocolate <img draggable="false" title=":heart_eyes:" alt="😍" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="twemoji" style="height: 1.3em;" width="20">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", class_name: 'twemoji', img_attrs: { style: 'height: 1.3em;', width: "20" })
   end
@@ -117,64 +140,19 @@ class TwemojiTest < Minitest::Test
       alt:   shortname_filter,
     }
 
-    expected = %(I like chocolate <img draggable="false" title="heart_eyes" alt="heart_eyes" src="https://twemoji.maxcdn.com/16x16/1f60d.png" class="twemoji" style="height: 1.3em;">!)
+    expected = %(I like chocolate <img draggable="false" title="heart_eyes" alt="heart_eyes" src="https://twemoji.maxcdn.com/2/svg/1f60d.svg" class="twemoji" style="height: 1.3em;">!)
 
     assert_equal expected, Twemoji.parse("I like chocolate :heart_eyes:!", class_name: 'twemoji', img_attrs: img_attrs)
   end
 
+  def test_parse_diversity_emojis
+    DiversityEmoji.all.each do |diversity|
+      # Each diversity set of emojis (base + 5 modifier = 6) should result in 6 img tags
+      assert_equal 6, Twemoji.parse(diversity).scan(/<img/).size
+    end
+  end
+
   def test_emoji_pattern
     assert_kind_of Regexp, Twemoji.emoji_pattern
-  end
-
-  def test_to_json_unsupported_file_ext
-    assert_raises RuntimeError do
-      Twemoji.to_json(file_ext: "bmp")
-    end
-  end
-
-  def test_to_json_unsupported_png_size
-    assert_raises RuntimeError do
-      Twemoji.to_json(file_ext: "png", image_size: "24x24")
-    end
-  end
-
-  def test_to_json_svg_format
-    json = assert_nothing_raised do
-      Twemoji.to_json(file_ext: "svg")
-    end
-
-    refute_includes json, "png"
-  end
-
-  def test_to_json_svg_format_ignore_image_size
-    json = assert_nothing_raised do
-      Twemoji.to_json(file_ext: "svg", image_size: "16x16")
-    end
-
-    refute_includes json, "png"
-  end
-
-  def test_to_json_png_format_size_16x16
-    json = assert_nothing_raised do
-      Twemoji.to_json(file_ext: "png", image_size: "16x16")
-    end
-
-    refute_includes json, "svg"
-  end
-
-  def test_to_json_png_format_size_36x36
-    json = assert_nothing_raised do
-      Twemoji.to_json(file_ext: "png", image_size: "36x36")
-    end
-
-    refute_includes json, "svg"
-  end
-
-  def test_to_json_png_format_size_72x72
-    json = assert_nothing_raised do
-      Twemoji.to_json(file_ext: "png", image_size: "72x72")
-    end
-
-    refute_includes json, "svg"
   end
 end